CLN: add formats/printing.py, strip from common.py
